# Add basic pure benchmarks