--- a/random.cabal
+++ b/random.cabal
@@ -28,9 +28,11 @@ custom-setup
         cabal-doctest >=1.0.6 && <1.1
 
 library
-    exposed-modules:  System.Random
-                    , System.Random.Internal
-                    , System.Random.Monad
+    exposed-modules:
+        System.Random
+        System.Random.Internal
+        System.Random.Monad
+
     c-sources:        cbits/CastFloatWord.cmm
     default-language: Haskell2010
     ghc-options:      -Wall
@@ -102,3 +104,14 @@ benchmark legacy-bench
         rdtsc -any,
         split >=0.2 && <0.3,
         time >=1.8 && <1.11
+
+benchmark bench
+    type:           exitcode-stdio-1.0
+    main-is:        Main.hs
+    hs-source-dirs: bench
+    ghc-options:    -Wall -O2
+    build-depends:
+        base >=4.10 && <5,
+        gauge >=0.2.3 && <0.3,
+        random -any,
+        splitmix >=0.0.3 && <0.1