@idontgetoutmuch 2020-04-22
I'll put these in the changelog rather than the old school benchmarks.