@curiousleo 2020-04-22
@idontgetoutmuch wrote:
> I'll put these in the changelog rather than the old school benchmarks.

Fantastic!

I've backported the benchmarks in their current state to `master` here: https://github.com/idontgetoutmuch/random/pull/111.

Those results are valid to compare to the ones posted here - same machine, same benchmarking code, etc. I had to benchmark with `--quick` because they took so long ... this means the accuracy is not as good, but this is about orders of magnitude anyway.