@curiousleo 2020-04-23
I didn't say it in the description originally, but I just want to make it clear that the times in the description are for 1048576 iterations in each case. I just stole that number from @lehins' benchmarks. I'm thinking of changing it to 1000000 just to make it a little easier to deduce the single iteration runtime.