@lehins 2020-04-27
@curiousleo `1048576` was no coincidence ;) I don't particularly care, it's just benchmarks, but just to give you a reason why `2^20=1048576` chosen. Thanks to this number you can easily translate from:
```
pure/uniform/Word8                       mean 293.3 μs
pure/uniform/Word16                      mean 293.6 μs
```
to it took `293.3 μs` to generate 1MiB of data of `Word8` type and it took `293.6 μs` for `2MiB` of data of `Word16`. I don't think anyone actually thinks in terms of a time it took to generate a single value, since that number is so low in these benchmarks.