@curiousleo 2020-04-28
@lehins ah cool, thanks for the explanation.

> I don't think anyone actually thinks in terms of a time it took to generate a single value, since that number is so low in these benchmarks.

It is _now_ :)

Good point though. I've actually reduces the number of iterations further from 1_000_000 to 100_000 in a follow-up PR so we can easily compare `uniformR` results as well, which takes longer.

It's not easy to communicate what these numbers mean. For `Word8` etc. it makes sense to talk about "generated data". But does it make sense to talk about "4MiB of Int32s in some range"?

Anyway, I'm just thinking aloud here. Thanks for your explanation.