@idontgetoutmuch 2020-04-28
I was about to put the new benchmarks in the CHANGELOG but we have regressions

```
pure/uniformR/full/Word16  0.017675  0.000026  67,528%
pure/uniformR/full/Int16  0.019081  0.030798     -38%
```

There really should *not* be a regression though?